requirements.txt
fastapi==0.133.1
uvicorn==0.41.0

numpy
pandas
scikit-learn

gensim==4.4.0
nltk==3.9.3
pymorphy3==2.0.6
razdel==0.5.0

openai==2.28.0
groq==1.0.0

python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.3

spacy==3.8.11

boto3==1.42.63
jieba==0.42.1

pytest==9.0.2
fonttools>=4.0.0

langchain==1.2.13
langchain-core==1.2.23
langchain-ollama==1.0.1

torch==2.2.2+cu126
--index-url https://download.pytorch.org/whl/cu126